Separate logic: public admin and dashboard
Public Admin app: logic to make Django Adminpublic and loginless

Dashboard app: logic for the views of Jarbas Admin
