+++ jarbas/public_admin/tests/test_dummy_user.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from jarbas.dashboard.sites import DummyUser
+from jarbas.public_admin.sites import DummyUser
 
 
 class TestDummyUser(TestCase):
